+++ CHANGELOG.md
@@ -24,3 +24,4 @@ For information on changes in released versions of Teku, see the [releases page]
  - Increase the batch size when searching for unknown validator indexes from 10 to 50.
  - Fixed issue with the voluntary-exit subcommand and Altair networks which caused "Failed to retrieve network config" errors.
  - Fixed issue where redundant attestations were incorrectly included in blocks.
+ - Validator performance is no longer logged when there are no attestations expected.